Track Auto-Rotation
Horizontal track pieces and corner pieces now automatically rotate
when
placed adjacent to other horizontal/corner piece tracks. If a
corner
track is placed adjacent to a single horizontal track, it will
rotate to
favor a right-hand turn, such that one click from the Hammer
in Rotate
Mode will make it a left-hand turn. If a corner piece is
placed adjacent
to two horizontal pieces at a 90 degree angle, it will
rotate to the
appropriate direction for the adjacent tracks.