Some Improvements to Slopes
Slope Up track pieces now place a dummy block when added to the world.
The function of this block is to simply be a hitbox to make placing the
next slope piece easier and more intuitive. Slope and Slope Down track
pieces do not yet have this functionality, but will in the near future,
I wanted to get this onto github. The Rotation Mode of the Hammer was
also modified to accomodate the new dummy block. When rotating a Slope
Up track piece, the Hammer tells the track to break the dummy, rotate,
then replace the dummy in a new location. Known Issues:
1: If two Slope Up track pieces are rotated/positioned such that their
dummy blocks overlap, the dummy of one piece will not exist after the
other rotates, this is not much of an issue, as all it does is remove
the convenience, the Slope Up track with the missing dummy simply needs
to be replaced. This will be fixed in the next push.
2: If a Slope Up track block is adjacent to any block and is rotated to
face the block, the block will be destroyed when the slope places it's
dummy again. This will be fixed in the next push.
3: The hammer overlay causes the player UI to pull from the wrong
texture sheet when looking at a track in survival (the hearts, hunger,
and experience).
4: All types of slopes do not currently have auto-rotate functionality.
I'm trying to refine the dummy placement mechanics first.

As a result of implementing this functionality, I am further tempted to
make a near-complete rewrite of the BlockTrack class, though I do not
know the exact details of how things would be laid out.