Tracks now use ForgeDirection
BlockTrack and TileEntityTrack now use ForgeDirection to determine which
way to face. Many things are broken, most notably the track rendering...
and the rotation with the hammer. This push is to let everyone take a
look at how it works, and maybe fix the rendering issues that I've
created by making the change.