Fixes yay.
Tracks now rotate on placement properly. Items temporarily set to use
their icons again, as the metadata switch broke their rendering. Started
some testing with loading .obj models, the latest test can be found in
the TrackPieceHorizontal class, to make the original block render again,
uncomment everything having to do with "model.renderAll()" and to
disable the .obj model from loading comment everything to do with
"test.renderAll()". Hammer was also fixed, and now properly rotates and
changes the states of the tracks, though the way in which style and
color are handled will need to be moved to the tile entity.