First tests of improved models
JSONHandler in the midst of a rewrite, before launching the mod, delete
everything inside of the "rollercoaster/track-styles/" folder, as the
model loader is having issues with reloading changed model files. New
Model Format:
-Each style will be defined as such:
1: folder of the style name in the "assets/rc/models/track-styles/"
folder
2: inside that folder are: A: the .json that defines the track pieces
for the model (make sure Model Format has a "." in front of it...)  and
B: the .obj or .tcn file. Both the .json AND the .tcn/.obj MUST have the
same name as their parent folder, this name is what becomes the in-code
reference call to the style and the model rendering.
3: inside of the single .tcn/.obj file will be multiple different
models, split up into "parts" or "groups." These MUST have the same name
as the parts that they are applied to, (ie. the model for horizontal
tracks must be named "horizontal" in the model file), this is how the
pieces themselves know what part of the model to render.

TODO:
-Better lighting for track models, the .objs I'm making will look so
much better if they're lit better.
-Fix the updating model file thing, (basically delete the directory or
always copy the files over).
-Adjust scale of models.
-Create models for the other track styles, currently only the
"corkscrew" style has the .obj single-file format.
-Set up the "place track as dummy" functionality for sloped tracks,
maybe come up with a way of having all models bigger than 1 block place
enough dummies to give the entire model hit boxes.

I'm excited to get new models in, and to start brainstorming new
pieces/mechanics and to start fleshing-out the TrackDesigner, so that we
can get this mod closer to what it should be. :P