A lot of stuff
-Deleted old BlockTrack and TileEntityTrack classes.
-Deleted old BlockDummy and TileEntityDummy classes.
-Moved direction handling to metadata, as well as dummy flags. If a
track block is a dummy, its metadata will be 10 higher than if it is a
normal track. 2 = North, 3 = South, 4 = West, 5 = East, 12 = Dummy
North, 13 = Dummy South, 14 = Dummy West, 15 = Dummy East. Dummies do
not render their block models.
-All attributes of tracks except for their type has been moved to the
TileEntityTrackBase class, and must be handled by the TE unless it is an
attribute that should apply to all instances of the track piece in
question.
-Fixed the new rendering (mostly). Loops and Heartline Rolls need
adjustment/different models, this will be adressed later.
-Tracks now place and rotate dummies as they should, this was made a lot
easier by the BlockSnapshot class in Forge 1230, as well as the new
PlaceEvent and MultiPlaceEvent events in BlockEvent. Sloped tracks
(being the only ones with dummies atm) are now smarter with their
placement mechanics, and will not allow you to place the block if the
location in which the dummy block should be placed is anything but air,
and the same goes for rotating, not allowing the rotation to occur if
the destination of the dummy (or the track itself - yes, sloped track
can be rotated from the dummy block now) is occupied by any thing but
air.
-Hammer now calls "rotate()" in the TileEntityTrackBase class, which
handles rotation completely on its own.

If I missed any bugs, let me know. This should fix 90% of the ones that
existed before this push.