A lot of stuff (again)
Changes:
-Right clicking on BlockFooter (or a BlockSupport that has a footer set)
with BlockSupport will add a support block to the top of the stack of
supports, until the stack is either 64 blocks tall (65 including the
footer), or until there is any block except air found above the top most
support. Breaking the footer also breaks all supports above it, as does
breaking any support in the stack
-Pick Blocking tracks now gives the proper track item
-right clicking on a painted track block with a water bucket will remove
the color from it and set the color back to default
-right clicking on a painted support or footer with a water bucket will
remove the color from the entire stack of supports at once
-right clicking a support or footer with a paint brush will paint the
entire stack of supports at once, shift right clicking will only paint
the support or footer that was clicked on
-right clicking a support or footer with a water bucket will clear the
paint from the entire column of connected supports, shift right clicking
will clear only the block that was clicked on
-tracks are now registered as ItemBlocks, this was done to assist in
fixing the pick block bug (the player was getting unstackable items with
different damage values)
-rendering for TileEntitySupport was adjusted to make code cleaner
-rendering for TrackPieces has been changed. instead of having 2
different "kinds" of track render functions, we now only have
"renderTileEntity" and "renderItem" instead of having both of those plus
"renderSpecialTileEntity" and "renderSpecialItem," the getter methods
for the render offsets have also been consolidated to simply
"getX/getY/getZ," the same applies to track extras as well
-the Hammer has been modified, shift right clicking on a track block now
rotates the track counter-clockwise, and does not change the hammer's
mode. to change the hammer's mode, the player must shift right click on
either a non-track block, or in the air
-new debug command, /trackitemanimation or /tia to disable the downward
motion of the held item in first person when switching between items,
this command isn't actually registered in the push, but can be by
uncommenting the line in RCMod
-more text added to GuiHammerOverlay
-track designer now has its own keybinding window, and fully rebindable
keys
-started implementing balloons, at this point they are only an item that
doesn't do anything... plan in the future is to make a custom entity
that "follows" the player on a line when the player moves, like an
actual balloon
-anything else i forgot about... :P

I think I might start porting more of the code base to 1.8, the most
tedious part will likely be the obj models... the new json system is
crazy. We need to discuss when we should release v 1.5, its been a
while.