Updating Forge...
Still not sure why I need to push the build.gradle everytime I want to
update forge but whatever. Next push will have some changes to the way
that dummies for tracks are handled, this is to make having more than 1
dummy per track easier. As a result, the Multi-Block system used by the
Track Fabricator (and soon the tracks) is receiving an update that will
help with rotation of structures.