Pushing to Sync
I've been working on further improving tracks, including support for
more than one "dummy" track block (which involved rewriting a good
portion of the multiblock system), switched TrackPiece and TrackExtra
registrations to an enum inside of TrackHandler, to make comparing with
track blocks less trivial, and not based on spelling. Added an
additional block icon to Track Fabricator Glass, the top/bottom faces
are now square textures instead of rectangles. Deleted the 1.8 block
model jsons as they are not needed in this branch. I am currently
working on making the roller coasters rideable again, (and also learning
how minecart physics work along the way), it is because of this that
there are 2 train items, in the future this won't be the case. Removed
large blocks of commented out code that have since been replaced with
better/functioning code.

Happy New Year, and hopefully we can get a release version out soon. :D