Some Changes to Stall GUIs
IMPORTANT NOTE:  We are now using an Access Transformer for disabling
the item animations for track blocks when debugging. To set it up in the
dev environment, run "gradlew clean", "gradlew setupDecompWorkspace",
and "gradlew --refresh-dependencies". If you're in IDEA you'll have to
also press the Syncronize button in the gradle panel. After that is
done, the 3 fields in the RCTickHandler that were marked private in
ItemRenderer will be public.

Stall GUIs now have a button on the top right that will switch the gui
to "Camouflage" mode, in which a player can place any block to have it's
texture be used as the block's texture. The GUI is a bit buggy atm as I
have not figured out how to handle the switching of Containers yet. We
also now have our own button class, that is able to make buttons
taller/shorter than vanilla ones.

Kyle, if you could look into making the tracks rideable again, that
would be great, I'm not experienced with entities. I will refine/finish
making the rest of the track models, and finish the new dummy handling
for more than one at a time.

Over the last few days I have been working with another modder in the
#minecraftforge irc to refine a .obj model loader for 1.8, and it is
currently in a functioning state, though there are a couple bugs that
still need to be squashed, but it looks like we'll be able to use .obj
files in 1.8! Also, this new loader has -limited- support for .mtl
files, which will make adding textures for .obj models easier. We'll
still need .json files for the item models, as the "camera transforms"
as they are called are handled through the vanilla system.